@@ -1,2 +1,4 @@
+# frozen_string_literal: true
+
 require 'csv_class_maker/csv_class_maker'
-require "csv_find/csv_find"
+require 'csv_find/csv_find'
