@@ -1,14 +1,16 @@
+# frozen_string_literal: true
+
 # Usage:
 #
 # require './csv_class_maker'
 # CsvClassMaker::generate_class 'Batting', '/Users/mplatt/Desktop/Batting.csv'
 
-module CsvClassMaker
+module CsvClassMaker #:nodoc:
   require 'csv'
 
   VERSION = {
-    major: '1',
-    minor: '1',
+    major: '2',
+    minor: '0',
     patch: '0'
   }.values.join('.')
 
@@ -16,7 +18,6 @@ module CsvClassMaker
     Object.const_set(class_name, Class.new do
       include CsvFind
       csv_file(file_name, options)
-    end
-    )
+    end)
   end
 end
